+++ ShellIiPanels.qml
@@ -23,6 +23,7 @@ import qs.modules.sidebarRight
 import qs.modules.tilingOverlay
 import qs.modules.verticalBar
 import qs.modules.wallpaperSelector
+import qs.modules.wallpaperLauncher
 import qs.modules.ii.overlay
 import qs.modules.shellUpdate
 import qs.modules.workspaceStrip
@@ -214,6 +215,7 @@ Item {
     }
     WallpaperSelectorRouter {}
     OnDemandPanelLoader { identifier: "iiWallpaperSelector"; open: GlobalStates.wallpaperSelectorOpen; retainAfterUse: true; closeGraceMs: 250; component: WallpaperSelector {} }
+    OnDemandPanelLoader { identifier: "iiWallpaperLauncher"; open: GlobalStates.wallpaperLauncherOpen; retainAfterUse: true; closeGraceMs: 250; component: WallpaperLauncher {} }
     OnDemandPanelLoader { identifier: "iiCoverflowSelector"; open: GlobalStates.coverflowSelectorOpen; retainAfterUse: true; closeGraceMs: 300; component: WallpaperCoverflow {} }
     OnDemandPanelLoader { identifier: "iiClipboard"; open: GlobalStates.clipboardOpen; retainAfterUse: true; closeGraceMs: 250; component: ClipboardModule.ClipboardPanel {} }
     OnDemandPanelLoader { identifier: "iiShellUpdate"; open: ShellUpdates.overlayOpen; closeGraceMs: 250; component: ShellUpdateOverlay {} }