+++ ShellWafflePanels.qml
@@ -11,6 +11,7 @@ import qs.modules.regionSelector
 import qs.modules.screenCorners
 import qs.modules.sessionScreen
 import qs.modules.wallpaperSelector
+import qs.modules.wallpaperLauncher
 import qs.modules.ii.overlay
 import qs.modules.workspaceStrip
 import "modules/clipboard" as ClipboardModule
@@ -115,6 +116,7 @@ Item {
     DeferredPanelLoader { identifier: "iiScreenCorners"; component: ScreenCorners {} }
     WallpaperSelectorRouter {}
     OnDemandPanelLoader { identifier: "iiWallpaperSelector"; open: GlobalStates.wallpaperSelectorOpen; retainAfterUse: true; closeGraceMs: 250; component: WallpaperSelector {} }
+    OnDemandPanelLoader { identifier: "iiWallpaperLauncher"; open: GlobalStates.wallpaperLauncherOpen; retainAfterUse: true; closeGraceMs: 250; component: WallpaperLauncher {} }
     OnDemandPanelLoader { identifier: "iiCoverflowSelector"; open: GlobalStates.coverflowSelectorOpen; retainAfterUse: true; closeGraceMs: 300; component: WallpaperCoverflow {} }
     DeferredPanelLoader { identifier: "iiClipboard"; extraCondition: Config.options?.panelFamily !== "waffle"; component: ClipboardModule.ClipboardPanel {} }
     OnDemandPanelLoader { identifier: "iiRecordingOsd"; open: RecorderStatus.isRecording; closeGraceMs: 250; component: RecordingOsd {} }